+++ CHANGES.rst
@@ -96,6 +96,8 @@ Other Changes and Additions
 
 - Subset selection dropdowns in plugins now show synced color indicators. [#1156, #1175]
 
+- Line analysis plugin now shows uncertainties, when available. [#1192]
+
 2.3 (2022-03-01)
 ================